chore: Extract ScaffoldNavBar from main.dart
