+++ mobile/lib/common/scaffold_with_nav_bar.dart
@@ -1,0 +1,61 @@
+import 'package:flutter/material.dart';
+import 'package:get_10101/common/app_bar_wrapper.dart';
+import 'package:get_10101/util/constants.dart';
+import 'package:get_10101/features/wallet/wallet_screen.dart';
+import 'package:get_10101/features/trade/trade_screen.dart';
+import 'package:go_router/go_router.dart';
+
+// Wrapper for the main application screens
+class ScaffoldWithNavBar extends StatelessWidget {
+  const ScaffoldWithNavBar({
+    required this.child,
+    Key? key,
+  }) : super(key: key);
+
+  final Widget child;
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      body: child,
+      appBar: const PreferredSize(
+          preferredSize: Size.fromHeight(40), child: SafeArea(child: AppBarWrapper())),
+      bottomNavigationBar: BottomNavigationBar(
+        items: <BottomNavigationBarItem>[
+          BottomNavigationBarItem(
+            icon: Container(key: tabWallet, child: const Icon(Icons.wallet)),
+            label: WalletScreen.label,
+          ),
+          BottomNavigationBarItem(
+            icon: Container(key: tabTrade, child: const Icon(Icons.bar_chart)),
+            label: TradeScreen.label,
+          ),
+        ],
+        currentIndex: _calculateSelectedIndex(context),
+        onTap: (int idx) => _onItemTapped(idx, context),
+      ),
+    );
+  }
+
+  static int _calculateSelectedIndex(BuildContext context) {
+    final String location = GoRouterState.of(context).location;
+    if (location.startsWith(WalletScreen.route)) {
+      return 0;
+    }
+    if (location.startsWith(TradeScreen.route)) {
+      return 1;
+    }
+    return 0;
+  }
+
+  void _onItemTapped(int index, BuildContext context) {
+    switch (index) {
+      case 0:
+        GoRouter.of(context).go(WalletScreen.route);
+        break;
+      case 1:
+        GoRouter.of(context).go(TradeScreen.route);
+        break;
+    }
+  }
+}
